+++ Snap Register System Interface/Snap Register System Interface/RegisterWindowParts/Business Objects/Sale.cs
@@ -4,14 +4,14 @@ namespace SnapRegisters
 {
     public class Sale: IDiscount
     {
-        public Sale( bool flat, string name, double amt )
+        public Sale( bool flat, string name, decimal amt )
         {
             Flat = flat;
             Name = name;
             Amount = amt;
         }
 
-        public double ChangeAmountTo( double amt )
+        public decimal ChangeAmountTo( decimal amt )
         {
             //Don't let the value go below 0 if it's a subtraction.
             return Flat ? Math.Max( amt - Amount, 0 ) : amt * ( 1 - Amount );
@@ -23,11 +23,11 @@ namespace SnapRegisters
         }
 
         public bool IsFlat() { return Flat; }
-        public double Discount() { return Amount; }
+        public decimal Discount() { return Amount; }
 
         public bool Flat { get; set; } = false;
         public string Name { get; set; } = string.Empty;
-        public double Amount { get; set; } = 0.0;
+        public decimal Amount { get; set; } = 0M;
 
     }
 }
